Policy add

+++ app/Http/Controllers/ArticleController.php
@@ -5,16 +5,16 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ArticleStoreRequest;
 use App\Http\Requests\ArticleUpdateRequest;
 use App\Models\Article;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
 
 class ArticleController extends ApiController implements HasMiddleware
 {
-    public static function middleware()
+    public static function middleware(): array
     {
         return [
-            new Middleware('auth:sanctum', except: ['index', 'show']),
+            new Middleware('auth:sanctum', except: ['index', 'show'])
         ];
     }
     /**
@@ -57,11 +57,15 @@ class ArticleController extends ApiController implements HasMiddleware
      */
     public function update(ArticleUpdateRequest $request, $id)
     {
+
         $article = Article::find($id);
 
         if (!$article) {
             return $this->errorResponse("Article not found.", 404);
         }
+
+        Gate::authorize('update', $article);
+
         $article->update($request->validated());
 
         return $this->successResponse($article, "Article updated successfully.");
@@ -76,6 +80,9 @@ class ArticleController extends ApiController implements HasMiddleware
         if (!$article) {
             return $this->errorResponse("Article not found.", 404);
         }
+
+        Gate::authorize('update', $article);
+
         $article->delete();
         return $this->successResponse(null, "Article deleted successfully.");
     }
